requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==0.5.0
Django==3.2.11
django-filter==23.1
django-floppyforms==1.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
mysqlclient==2.1.1
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
requests==2.28.2
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
urllib3==1.26.15
whitenoise==5.3.0
